rename: 패키지 이름 변경 #7
 github_oauth -> github 으로 패키지 이름 변경
