+++ be/src/main/java/team05a/secondhand/member/repository/MemberRepository.java
@@ -5,7 +5,7 @@ import java.util.Optional;
 import org.springframework.data.jpa.repository.JpaRepository;
 
 import team05a.secondhand.member.data.entity.Member;
-import team05a.secondhand.oauth_github.OauthAttributes;
+import team05a.secondhand.oauth.OauthAttributes;
 
 public interface MemberRepository extends JpaRepository<Member, Long> {
