+++ be/src/main/java/team05a/secondhand/oauth/data/dto/LoginResponse.java
@@ -1,4 +1,4 @@
-package team05a.secondhand.oauth_github.data.dto;
+package team05a.secondhand.oauth.data.dto;
 
 import java.util.List;
